Change search mode for Anilist lookups
The old search mode of "descending popularity" made some entries
unretrievable, even if one provided an exact match, due to that entry
not being the most popular result for the query.
By changing the search mode to "search match", such entries are now
preferred.
Additionally, this mirrors the lookup defaults for the Anilist web
interface, so there should be no more inconsistent results between
searching on the website or using the bot.