Use Anilist's "Mean" instead of "Average Score"
As discussed in
https://twitter.com/FallenWarrior2k/status/1045042888374194178,
Anilist's "Average Score" uses a weird "weighting"/"smoothing" formula
instead of counting every vote equally.
This might be useful for titles with single-digit amounts of votes, but
it artificially drags down "low profile"-ish series, as shown in the
Twitter thread.
As such, I've decided to use the "Mean Score" instead.