fix(saucenao): avoid ghost typing
Editing a message does not reset typing, so it will often linger after
the embed has already been posted. By not sending the initial search
message and editing that, but rather using just the typing to indicate
work and then sending a message with the result that automatically
cancels the typing, the whole thing should be "smoother".